refactor tests with ea framework utils
