--- packages/sources/cfbenchmarks/test/integration/setup.ts
@@ -1,103 +1,0 @@
-import { ServerInstance } from '@chainlink/external-adapter-framework'
-import { Adapter } from '@chainlink/external-adapter-framework/adapter'
-import { WebSocketClassProvider } from '@chainlink/external-adapter-framework/transports'
-import { Server, WebSocket } from 'mock-socket'
-import * as process from 'process'
-import { SuperTest, Test } from 'supertest'
-import { config } from '../../src/config'
-import { birc, crypto, cryptolwba } from '../../src/endpoint'
-
-export type SuiteContext = {
-  req: SuperTest<Test> | null
-  server: () => Promise<ServerInstance>
-  fastify?: ServerInstance
-}
-
-export type EnvVariables = { [key: string]: string }
-
-export type TestOptions = { cleanNock?: boolean; fastify?: boolean }
-
-/**
- * Sets the mocked websocket instance in the provided provider class.
- * We need this here, because the tests will connect using their instance of WebSocketClassProvider;
- * fetching from this library to the \@chainlink/ea-bootstrap package would access _another_ instance
- * of the same constructor. Although it should be a singleton, dependencies are different so that
- * means that the static classes themselves are also different.
- *
- * @param provider - singleton WebSocketClassProvider
- */
-export const mockWebSocketProvider = (provider: typeof WebSocketClassProvider): void => {
-  // Extend mock WebSocket class to bypass protocol headers error
-  class MockWebSocket extends WebSocket {
-    constructor(url: string, protocol: string | string[] | Record<string, string> | undefined) {
-      super(url, protocol instanceof Object ? undefined : protocol)
-    }
-    // This is part of the 'ws' node library but not the common interface, but it's used in our WS transport
-    removeAllListeners() {
-      for (const eventType in this.listeners) {
-        // We have to manually check because the mock-socket library shares this instance, and adds the server listeners to the same obj
-        if (!eventType.startsWith('server')) {
-          delete this.listeners[eventType]
-        }
-      }
-    }
-  }
-
-  // Need to disable typing, the mock-socket impl does not implement the ws interface fully
-  provider.set(MockWebSocket as any) // eslint-disable-line @typescript-eslint/no-explicit-any
-}
-
-export const mockCryptoWebSocketServer = (URL: string): Server => {
-  const mockWsServer = new Server(URL, { mock: false })
-  mockWsServer.on('connection', (socket) => {
-    socket.on('message', (_) => {
-      socket.send(
-        JSON.stringify({
-          type: 'value',
-          time: 1645203822000,
-          id: 'BRTI',
-          value: '40067.00',
-        }),
-      )
-    })
-  })
-
-  return mockWsServer
-}
-
-export const mockLwbaWebSocketServer = (URL: string): Server => {
-  const mockWsServer = new Server(URL, { mock: false })
-  mockWsServer.on('connection', (socket) => {
-    socket.on('message', (_) => {
-      socket.send(
-        JSON.stringify({
-          type: 'rti_stats',
-          time: 1677876163000,
-          id: 'U_ETHUSD_RTI',
-          value: '1.1635',
-          utilizedDepth: '1888000.0',
-          valueAsk: '1.1662',
-          valueBid: '1.1607',
-          midPrice: '1.1631',
-        }),
-      )
-    })
-  })
-
-  return mockWsServer
-}
-
-export const createAdapter = () => {
-  return new Adapter({
-    name: 'CFBENCHMARKS',
-    endpoints: [crypto, birc, cryptolwba],
-    defaultEndpoint: crypto.name,
-    config,
-  })
-}
-
-export function setEnvVariables(envVariables: NodeJS.ProcessEnv): void {
-  for (const key in envVariables) {
-    process.env[key] = envVariables[key]
-  }
-}
