+++ packages/sources/coingecko/test/integration/adapter.test.ts
@@ -1,66 +1,49 @@
 import { mockCryptoSuccess, mockDominanceSuccess } from './fixtures'
-import { SuperTest, Test } from 'supertest'
-import { setupExternalAdapterTest, SuiteContext } from './setup'
-import { ServerInstance } from '@chainlink/external-adapter-framework'
+import {
+  TestAdapter,
+  setEnvVariables,
+} from '@chainlink/external-adapter-framework/util/testing-utils'
+import * as nock from 'nock'
 
 describe('execute', () => {
   let spy: jest.SpyInstance
+  let testAdapter: TestAdapter
+  let oldEnv: NodeJS.ProcessEnv
+
   beforeAll(async () => {
+    oldEnv = JSON.parse(JSON.stringify(process.env))
     const mockDate = new Date('2022-05-10T16:09:27.193Z')
     spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
+
+    const adapter = (await import('./../../src')).adapter
+    adapter.rateLimiting = undefined
+    testAdapter = await TestAdapter.startWithMockedCache(adapter, {
+      testAdapter: {} as TestAdapter<never>,
+    })
   })
 
-  afterAll((done) => {
+  afterAll(async () => {
+    setEnvVariables(oldEnv)
+    await testAdapter.api.close()
+    nock.restore()
+    nock.cleanAll()
     spy.mockRestore()
-    done()
   })
-
-  const id = '1'
-
-  const context: SuiteContext = {
-    req: null,
-    server: async () => {
-      process.env['RATE_LIMIT_CAPACITY_SECOND'] = '5000'
-      process.env['CACHE_POLLING_MAX_RETRIES'] = '20'
-      process.env['CACHE_POLLING_SLEEP_MS'] = '500'
-      process.env['METRICS_ENABLED'] = 'false'
-      process.env['BACKGROUND_EXECUTE_MS_HTTP'] = '50'
-      const server = (await import('../../src')).server
-      return server() as Promise<ServerInstance>
-    },
-  }
-
-  const envVariables = {
-    CACHE_ENABLED: 'false',
-  }
-
-  setupExternalAdapterTest(envVariables, context)
 
   describe('crypto api', () => {
-    const data = {
-      id,
-      data: {
+    it('should return success', async () => {
+      const data = {
         base: 'ETH',
         quote: 'USD',
-      },
-    }
-
-    it('should return success', async () => {
+      }
       mockCryptoSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
 
-    const dataWithOverride = {
-      id,
-      data: {
+    it('should return success for override', async () => {
+      const data = {
         base: 'OHMV2',
         quote: 'USD',
         overrides: {
@@ -68,25 +51,15 @@ describe('execute', () => {
             OHMV2: 'olympus',
           },
         },
-      },
-    }
-
-    it('should return success for override', async () => {
+      }
       mockCryptoSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(dataWithOverride)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
 
-    const dataWithArray = {
-      id,
-      data: {
+    it('should return failure for array', async () => {
+      const data = {
         base: ['OHMV2', 'ETH'],
         quote: 'USD',
         overrides: {
@@ -94,114 +67,65 @@ describe('execute', () => {
             OHMV2: 'olympus',
           },
         },
-      },
-    }
-
-    it('should return failure for array', async () => {
+      }
       mockCryptoSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(dataWithArray)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(400)
-      expect(response.body).toMatchSnapshot()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(400)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 
   describe('volume api', () => {
-    const data = {
-      id,
-      data: {
+    it('should return success', async () => {
+      const data = {
         endpoint: 'volume',
         base: 'ETH',
         quote: 'USD',
-      },
-    }
-
-    it('should return success', async () => {
+      }
       mockCryptoSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 
   describe('marketcap api', () => {
-    const data = {
-      id,
-      data: {
+    it('should return success', async () => {
+      const data = {
         endpoint: 'marketcap',
         base: 'ETH',
         quote: 'USD',
-      },
-    }
-
-    it('should return success', async () => {
+      }
       mockCryptoSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 
   describe('globalmarketcap api', () => {
-    const data = {
-      id,
-      data: {
+    it('should return success', async () => {
+      const data = {
         endpoint: 'globalmarketcap',
         quote: 'USD',
-      },
-    }
-
-    it('should return success', async () => {
+      }
       mockDominanceSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 
   describe('dominance api', () => {
-    const data = {
-      id,
-      data: {
+    it('should return success', async () => {
+      const data = {
         endpoint: 'dominance',
         quote: 'ETH',
-      },
-    }
-
-    it('should return success', async () => {
+      }
       mockDominanceSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 })
