--- packages/sources/coinmarketcap/test/integration/setup.ts
@@ -1,56 +1,0 @@
-import request, { SuperTest, Test } from 'supertest'
-import { AddressInfo } from 'net'
-import * as process from 'process'
-import * as nock from 'nock'
-import { ServerInstance } from '@chainlink/external-adapter-framework'
-
-export type SuiteContext = {
-  req: SuperTest<Test> | null
-  server: () => Promise<ServerInstance>
-  fastify?: ServerInstance
-}
-
-export type EnvVariables = { [key: string]: string }
-
-export type TestOptions = { cleanNock?: boolean; fastify?: boolean }
-
-export const setupExternalAdapterTest = (
-  envVariables: NodeJS.ProcessEnv,
-  context: SuiteContext,
-  options: TestOptions = { cleanNock: true, fastify: false },
-): void => {
-  let fastify: ServerInstance
-
-  beforeAll(async () => {
-    process.env['METRICS_ENABLED'] = 'false'
-    for (const key in envVariables) {
-      process.env[key] = envVariables[key]
-    }
-
-    if (process.env['RECORD']) {
-      nock.recorder.rec()
-    }
-    fastify = await context.server()
-
-    // eslint-disable-next-line require-atomic-updates
-    context.req = request(`localhost:${(fastify.server.address() as AddressInfo).port}`)
-
-    // Only for edge cases when someone needs to use the fastify instance outside this function
-    if (options.fastify) {
-      // eslint-disable-next-line require-atomic-updates
-      context.fastify = fastify
-    }
-  })
-
-  afterAll(async () => {
-    if (process.env['RECORD']) {
-      nock.recorder.play()
-    }
-
-    nock.restore()
-    nock.cleanAll()
-    nock.enableNetConnect()
-
-    await fastify.close()
-  })
-}
