+++ packages/sources/coinmetrics/test/integration/adapter-ws.test.ts
@@ -1,138 +1,116 @@
-import * as process from 'process'
-import { AddressInfo } from 'net'
-import { AdapterRequestBody, sleep } from '@chainlink/external-adapter-framework/util'
-import {
-  createAdapter,
-  mockWebSocketProvider,
-  mockWebSocketServer,
-  setEnvVariables,
-} from './setup-ws'
-import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
-import request, { SuperTest, Test } from 'supertest'
+import { TestAdapter, setEnvVariables } from '@chainlink/external-adapter-framework/util/test-util'
 import { WebSocketClassProvider } from '@chainlink/external-adapter-framework/transports'
 import { Server } from 'mock-socket'
+import {
+  mockWebSocketServer,
+  mockWebSocketProvider,
+  mockCryptoLwbaWebSocketServer,
+} from './fixtures'
+import FakeTimers from '@sinonjs/fake-timers'
 
 describe('websocket', () => {
-  let fastify: ServerInstance | undefined
-  let req: SuperTest<Test>
   let mockWsServer: Server | undefined
-  let spy: jest.SpyInstance
+  let mockWsServerLwba: Server | undefined
+  let testAdapter: TestAdapter
+  let oldEnv: NodeJS.ProcessEnv
   const wsEndpoint = 'ws://localhost:9090/v4/timeseries-stream/asset-metrics'
+  const wsEndpointLwba = 'ws://localhost:9090/v4/timeseries-stream/pair-quotes'
 
-  jest.setTimeout(30_000)
-
-  const data: AdapterRequestBody = {
-    data: {
-      base: 'ETH',
-      quote: 'USD',
-    },
+  const data = {
+    base: 'ETH',
+    quote: 'USD',
   }
 
-  let oldEnv: NodeJS.ProcessEnv
+  const dataLwba = {
+    endpoint: 'crypto-lwba',
+    base: 'ETH',
+    quote: 'USD',
+  }
+
   beforeAll(async () => {
     oldEnv = JSON.parse(JSON.stringify(process.env))
     process.env['WS_SUBSCRIPTION_TTL'] = '5000'
     process.env['CACHE_MAX_AGE'] = '5000'
     process.env['CACHE_POLLING_MAX_RETRIES'] = '0'
-    process.env['METRICS_ENABLED'] = 'false'
     process.env['WS_API_ENDPOINT'] = wsEndpoint
     process.env['API_KEY'] = 'fake-api-key'
 
     // Start mock web socket server
     mockWebSocketProvider(WebSocketClassProvider)
     mockWsServer = mockWebSocketServer(wsEndpoint)
+    mockWsServerLwba = mockCryptoLwbaWebSocketServer(wsEndpointLwba)
 
-    const mockDate = new Date('2022-05-10T16:09:27.193Z')
-    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
+    const adapter = (await import('./../../src')).adapter
+    testAdapter = await TestAdapter.startWithMockedCache(adapter, {
+      clock: FakeTimers.install(),
+      testAdapter: {} as TestAdapter<never>,
+    })
 
-    fastify = await expose(createAdapter())
-    req = request(`http://localhost:${(fastify?.server.address() as AddressInfo).port}`)
-
-    // Send initial request to start background execute
-    await req.post('/').send(data)
-    await sleep(5000)
+    // Send initial request to start background execute and wait for cache to be filled with results
+    await testAdapter.request(data)
+    await testAdapter.request(dataLwba)
+    await testAdapter.waitForCache(2)
   })
 
-  afterAll((done) => {
-    spy.mockRestore()
+  afterAll(async () => {
     setEnvVariables(oldEnv)
     mockWsServer?.close()
-    fastify?.close(done())
+    mockWsServerLwba?.close()
+    testAdapter.clock?.uninstall()
+    await testAdapter.api.close()
   })
 
-  describe('websocket endpoint', () => {
+  describe('price endpoint', () => {
     it('should return success', async () => {
-      const makeRequest = () =>
-        req
-          .post('/')
-          .send(data)
-          .set('Accept', '*/*')
-          .set('Content-Type', 'application/json')
-          .expect('Content-Type', /json/)
-          .expect(200)
+      const response = await testAdapter.request(data)
+      expect(response.json()).toMatchSnapshot()
+    })
 
-      const response = await makeRequest()
-      expect(response.body).toMatchSnapshot()
-    }, 30000)
-    it('should return error (empty body)', async () => {
-      const makeRequest = () =>
-        req
-          .post('/')
-          .send({})
-          .set('Accept', '*/*')
-          .set('Content-Type', 'application/json')
-          .expect('Content-Type', /json/)
-
-      const response = await makeRequest()
-      expect(response.statusCode).toEqual(400)
-    }, 30000)
     it('should return error (empty data)', async () => {
-      const makeRequest = () =>
-        req
-          .post('/')
-          .send({ data: {} })
-          .set('Accept', '*/*')
-          .set('Content-Type', 'application/json')
-          .expect('Content-Type', /json/)
-
-      const response = await makeRequest()
+      const response = await testAdapter.request({})
       expect(response.statusCode).toEqual(400)
-    }, 30000)
+    })
+
     it('should return error (empty base)', async () => {
-      const makeRequest = () =>
-        req
-          .post('/')
-          .send({ data: { quote: 'USD' } })
-          .set('Accept', '*/*')
-          .set('Content-Type', 'application/json')
-          .expect('Content-Type', /json/)
-
-      const response = await makeRequest()
+      const response = await testAdapter.request({ quote: 'USD' })
       expect(response.statusCode).toEqual(400)
-    }, 30000)
+    })
+
     it('should return error (empty quote)', async () => {
-      const makeRequest = () =>
-        req
-          .post('/')
-          .send({ data: { base: 'ETH' } })
-          .set('Accept', '*/*')
-          .set('Content-Type', 'application/json')
-          .expect('Content-Type', /json/)
-
-      const response = await makeRequest()
+      const response = await testAdapter.request({ base: 'ETH' })
       expect(response.statusCode).toEqual(400)
-    }, 30000)
+    })
+
     it('should return error (bad quote)', async () => {
-      const makeRequest = () =>
-        req
-          .post('/')
-          .send({ data: { base: 'ETH', quote: 'INVALID' } })
-          .set('Accept', '*/*')
-          .set('Content-Type', 'application/json')
-          .expect('Content-Type', /json/)
-
-      const response = await makeRequest()
+      const response = await testAdapter.request({ base: 'ETH', quote: 'INVALID' })
       expect(response.statusCode).toEqual(400)
-    }, 30000)
+    })
+  })
+
+  describe('lwba endpoint', () => {
+    it('should return success', async () => {
+      const response = await testAdapter.request(dataLwba)
+      expect(response.json()).toMatchSnapshot()
+    })
+
+    it('should return error (empty body)', async () => {
+      const response = await testAdapter.request({})
+      expect(response.statusCode).toEqual(400)
+    })
+
+    it('should return error (empty data)', async () => {
+      const response = await testAdapter.request({ endpoint: 'crypto-lwba' })
+      expect(response.statusCode).toEqual(400)
+    })
+
+    it('should return error (empty base)', async () => {
+      const response = await testAdapter.request({ endpoint: 'crypto-lwba', quote: 'USD' })
+      expect(response.statusCode).toEqual(400)
+    })
+
+    it('should return error (empty quote)', async () => {
+      const response = await testAdapter.request({ endpoint: 'crypto-lwba', base: 'ETH' })
+      expect(response.statusCode).toEqual(400)
+    })
   })
 })
