+++ packages/sources/coinmetrics/test/integration/adapter.test.ts
@@ -1,105 +1,70 @@
-import { SuperTest, Test } from 'supertest'
-import { setupExternalAdapterTest, SuiteContext } from './setup'
-import { ServerInstance } from '@chainlink/external-adapter-framework'
 import { mockCoinmetricsResponseSuccess, mockCoinmetricsResponseSuccess2 } from './fixtures'
+import { TestAdapter, setEnvVariables } from '@chainlink/external-adapter-framework/util/test-util'
+import * as nock from 'nock'
 
 describe('http', () => {
   let spy: jest.SpyInstance
+  let testAdapter: TestAdapter
+  let oldEnv: NodeJS.ProcessEnv
+
   beforeAll(async () => {
+    oldEnv = JSON.parse(JSON.stringify(process.env))
+    process.env['API_KEY'] = 'fake-api-key'
     const mockDate = new Date('2022-01-01T11:11:11.111Z')
     spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
+
+    const adapter = (await import('./../../src')).adapter
+    adapter.rateLimiting = undefined
+    testAdapter = await TestAdapter.startWithMockedCache(adapter, {
+      testAdapter: {} as TestAdapter<never>,
+    })
   })
 
-  afterAll((done) => {
+  afterAll(async () => {
+    setEnvVariables(oldEnv)
+    await testAdapter.api.close()
+    nock.restore()
+    nock.cleanAll()
     spy.mockRestore()
-    done()
   })
-
-  const id = '1'
-
-  const context: SuiteContext = {
-    req: null,
-    server: async () => {
-      process.env['RATE_LIMIT_CAPACITY_SECOND'] = '6'
-      process.env['METRICS_ENABLED'] = 'false'
-      const server = (await import('../../src')).server
-      return server() as Promise<ServerInstance>
-    },
-  }
-
-  const envVariables = {
-    API_KEY: process.env.API_KEY || 'fake-api-key',
-  }
-
-  setupExternalAdapterTest(envVariables, context)
 
   describe('price endpoint', () => {
-    const data = {
-      id,
-      data: {
+    it('should return success', async () => {
+      const data = {
         base: 'BTC',
         quote: 'USD',
         transport: 'rest',
-      },
-    }
-
-    it('should return success', async () => {
+      }
       mockCoinmetricsResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 
   describe('burned endpoint', () => {
-    const data = {
-      id,
-      data: {
+    it('should return success', async () => {
+      const data = {
         endpoint: 'burned',
         asset: 'eth',
-      },
-    }
-
-    it('should return success', async () => {
+      }
       mockCoinmetricsResponseSuccess2()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 
   describe('total-burned endpoint', () => {
-    const data = {
-      id,
-      data: {
+    it('should return success', async () => {
+      const data = {
         endpoint: 'total-burned',
         asset: 'eth',
-      },
-    }
-
-    it('should return success', async () => {
+      }
       mockCoinmetricsResponseSuccess2(10000)
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 })
