+++ packages/sources/cryptocompare/package.json
@@ -32,8 +32,10 @@
     "tslib": "^2.3.1"
   },
   "devDependencies": {
+    "@sinonjs/fake-timers": "9.1.2",
     "@types/jest": "27.5.2",
     "@types/node": "16.11.51",
+    "@types/sinonjs__fake-timers": "8.1.2",
     "@types/supertest": "2.0.12",
     "mock-socket": "9.1.5",
     "nock": "13.2.9",
