+++ packages/sources/cryptocompare/test/integration/adapter-ws.test.ts
@@ -1,0 +1,72 @@
+import { mockWebSocketProvider, mockWebSocketServer } from './fixtures'
+import { Server } from 'mock-socket'
+import { WebSocketClassProvider } from '@chainlink/external-adapter-framework/transports'
+import FakeTimers from '@sinonjs/fake-timers'
+import { TestAdapter, setEnvVariables } from '@chainlink/external-adapter-framework/util/test-util'
+describe('websocket', () => {
+  let mockWsServer: Server | undefined
+  let testAdapter: TestAdapter
+  const wsEndpoint = 'ws://localhost:9090'
+  let oldEnv: NodeJS.ProcessEnv
+
+  const data = {
+    endpoint: 'crypto',
+    base: 'ETH',
+    quote: 'BTC',
+    transport: 'ws',
+  }
+
+  beforeAll(async () => {
+    oldEnv = JSON.parse(JSON.stringify(process.env))
+    process.env['WS_SUBSCRIPTION_TTL'] = '10000'
+    process.env['CACHE_MAX_AGE'] = '10000'
+    process.env['CACHE_POLLING_MAX_RETRIES'] = '0'
+    process.env['WS_API_ENDPOINT'] = wsEndpoint
+    process.env['API_KEY'] = 'fake-api-key'
+    process.env['WS_ENABLED'] = 'true'
+
+    // Start mock web socket server
+    mockWebSocketProvider(WebSocketClassProvider)
+    mockWsServer = mockWebSocketServer(wsEndpoint)
+
+    const adapter = (await import('./../../src')).adapter
+    adapter.rateLimiting = undefined
+    testAdapter = await TestAdapter.startWithMockedCache(adapter, {
+      clock: FakeTimers.install(),
+      testAdapter: {} as TestAdapter<never>,
+    })
+
+    // Send initial request to start background execute and wait for cache to be filled with result
+    await testAdapter.request(data)
+    await testAdapter.waitForCache()
+  })
+
+  afterAll(async () => {
+    setEnvVariables(oldEnv)
+    mockWsServer?.close()
+    testAdapter.clock?.uninstall()
+    await testAdapter.api.close()
+  })
+
+  describe('crypto endpoint', () => {
+    it('should return success', async () => {
+      const response = await testAdapter.request(data)
+      expect(response.json()).toMatchSnapshot()
+    })
+
+    it('should return error (empty data)', async () => {
+      const response = await testAdapter.request({})
+      expect(response.statusCode).toEqual(400)
+    })
+
+    it('should return error (empty base)', async () => {
+      const response = await testAdapter.request({ quote: 'BTC' })
+      expect(response.statusCode).toEqual(400)
+    })
+
+    it('should return error (empty quote)', async () => {
+      const response = await testAdapter.request({ base: 'ETH' })
+      expect(response.statusCode).toEqual(400)
+    })
+  })
+})
