+++ packages/sources/cryptocompare/test/integration/fixtures.ts
@@ -1,4 +1,6 @@
 import nock from 'nock'
+import { WebSocketClassProvider } from '@chainlink/external-adapter-framework/transports'
+import { Server, WebSocket } from 'mock-socket'
 
 export const mockCryptoSuccess = (): nock.Scope =>
   nock('https://min-api.cryptocompare.com', {
@@ -162,3 +164,52 @@ export const mockVwapSuccess = (): nock.Scope =>
       'Vary',
       'Origin',
     ])
+    .persist()
+
+export const mockWebSocketProvider = (provider: typeof WebSocketClassProvider): void => {
+  // Extend mock WebSocket class to bypass protocol headers error
+  class MockWebSocket extends WebSocket {
+    constructor(url: string, protocol: string | string[] | Record<string, string> | undefined) {
+      super(url, protocol instanceof Object ? undefined : protocol)
+    }
+    // This is part of the 'ws' node library but not the common interface, but it's used in our WS transport
+    removeAllListeners() {
+      for (const eventType in this.listeners) {
+        // We have to manually check because the mock-socket library shares this instance, and adds the server listeners to the same obj
+        if (!eventType.startsWith('server')) {
+          delete this.listeners[eventType]
+        }
+      }
+    }
+  }
+
+  // Need to disable typing, the mock-socket impl does not implement the ws interface fully
+  provider.set(MockWebSocket as any) // eslint-disable-line @typescript-eslint/no-explicit-any
+}
+
+const base = 'ETH'
+const quote = 'BTC'
+const price = 1234
+
+export const mockWebSocketServer = (URL: string) => {
+  const mockWsServer = new Server(URL, { mock: false })
+  mockWsServer.on('connection', (socket) => {
+    socket.send(
+      JSON.stringify({
+        MESSAGE: 'STREAMERWELCOME',
+      }),
+    )
+    const parseMessage = () => {
+      socket.send(
+        JSON.stringify({
+          TYPE: '5',
+          FROMSYMBOL: base,
+          TOSYMBOL: quote,
+          PRICE: price,
+        }),
+      )
+    }
+    socket.on('message', parseMessage)
+  })
+  return mockWsServer
+}
