+++ packages/sources/dxfeed/test/integration/adapter.test.ts
@@ -1,130 +1,42 @@
-import request, { SuperTest, Test } from 'supertest'
-import {
-  createAdapter,
-  mockWebSocketProvider,
-  mockWebSocketServer,
-  setEnvVariables,
-  setupExternalAdapterTest,
-  SuiteContext,
-} from './setup'
-import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
 import { mockPriceEndpoint } from './fixtures'
-import process from 'process'
-import { AddressInfo } from 'net'
-import { Server } from 'mock-socket'
-import { WebSocketClassProvider } from '@chainlink/external-adapter-framework/transports'
-import { sleep } from '@chainlink/external-adapter-framework/util'
+import { TestAdapter, setEnvVariables } from '@chainlink/external-adapter-framework/util/test-util'
+import * as nock from 'nock'
 
 describe('execute', () => {
-  describe('price endpoint rest', () => {
-    let spy: jest.SpyInstance
-    beforeAll(async () => {
-      const mockDate = new Date('2022-01-01T11:11:11.111Z')
-      spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
-    })
+  let spy: jest.SpyInstance
+  let testAdapter: TestAdapter
+  let oldEnv: NodeJS.ProcessEnv
+  beforeAll(async () => {
+    oldEnv = JSON.parse(JSON.stringify(process.env))
+    process.env['API_USERNAME'] = 'fake-api-username'
+    process.env['API_PASSWORD'] = 'fake-api-password'
+    const mockDate = new Date('2022-01-01T11:11:11.111Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
 
-    afterAll((done) => {
-      spy.mockRestore()
-      done()
-    })
-
-    const id = '1'
-
-    const context: SuiteContext = {
-      req: null,
-      server: async () => {
-        process.env['RATE_LIMIT_CAPACITY_SECOND'] = '6'
-        process.env['METRICS_ENABLED'] = 'false'
-        const server = (await import('../../src')).server
-        return server() as Promise<ServerInstance>
-      },
-    }
-
-    const envVariables = {
-      CACHE_ENABLED: 'false',
-      API_USERNAME: process.env.API_USERNAME || 'fake-api-username',
-      API_PASSWORD: process.env.API_PASSWORD || 'fake-api-password',
-    }
-
-    setupExternalAdapterTest(envVariables, context)
-
-    const data = {
-      id,
-      data: {
-        base: 'TSLA',
-      },
-    }
-
-    it('should return success', async () => {
-      mockPriceEndpoint()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+    const adapter = (await import('./../../src')).adapter
+    adapter.rateLimiting = undefined
+    testAdapter = await TestAdapter.startWithMockedCache(adapter, {
+      testAdapter: {} as TestAdapter<never>,
     })
   })
 
-  describe('price endpoint websocket', () => {
-    let fastify: ServerInstance | undefined
-    let req: SuperTest<Test>
-    let mockWsServer: Server | undefined
-    let spy: jest.SpyInstance
-    const wsEndpoint = 'ws://localhost:9090'
+  afterAll(async () => {
+    setEnvVariables(oldEnv)
+    await testAdapter.api.close()
+    nock.restore()
+    nock.cleanAll()
+    spy.mockRestore()
+  })
 
-    jest.setTimeout(100000)
-
-    const priceData = {
-      data: {
-        base: 'TSLA',
-        transport: 'ws',
-      },
-    }
-
-    let oldEnv: NodeJS.ProcessEnv
-    beforeAll(async () => {
-      oldEnv = JSON.parse(JSON.stringify(process.env))
-      process.env['WS_SUBSCRIPTION_TTL'] = '10000'
-      process.env['CACHE_MAX_AGE'] = '10000'
-      process.env['CACHE_POLLING_MAX_RETRIES'] = '0'
-      process.env['METRICS_ENABLED'] = 'false'
-      process.env['WS_API_ENDPOINT'] = wsEndpoint
-      const mockDate = new Date('2022-11-11T11:11:11.111Z')
-      spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
-
-      mockWebSocketProvider(WebSocketClassProvider)
-      mockWsServer = mockWebSocketServer(wsEndpoint)
-
-      fastify = await expose(createAdapter())
-      req = request(`http://localhost:${(fastify?.server.address() as AddressInfo).port}`)
-
-      // Send initial request to start background execute
-      await req.post('/').send(priceData)
-      await sleep(5000)
-    })
-
-    afterAll((done) => {
-      spy.mockRestore()
-      setEnvVariables(oldEnv)
-      mockWsServer?.close()
-      fastify?.close(done())
-    })
-
+  describe('price endpoint rest', () => {
     it('should return success', async () => {
-      const makeRequest = () =>
-        req
-          .post('/')
-          .send(priceData)
-          .set('Accept', '*/*')
-          .set('Content-Type', 'application/json')
-          .expect('Content-Type', /json/)
-
-      const response = await makeRequest()
-      expect(response.body).toMatchSnapshot()
-    }, 30000)
+      const data = {
+        base: 'TSLA',
+      }
+      mockPriceEndpoint()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
+    })
   })
 })
